+++ packages/brick-kit/src/ErrorBoundary.tsx
@@ -15,10 +15,10 @@ interface ErrorBoundaryState {
 
 // Ref https://reactjs.org/docs/error-boundaries.html
 class LegacyErrorBoundary extends React.Component<
-  WithTranslation,
+  WithTranslation<typeof NS_BRICK_KIT>,
   ErrorBoundaryState
 > {
-  constructor(props: WithTranslation) {
+  constructor(props: WithTranslation<typeof NS_BRICK_KIT>) {
     super(props);
     this.state = { error: null };
   }
@@ -49,6 +49,4 @@ class LegacyErrorBoundary extends React.Component<
 }
 
 /** @internal */
-export const ErrorBoundary: React.ComponentType<WithTranslationProps> = withTranslation(
-  NS_BRICK_KIT
-)(LegacyErrorBoundary);
+export const ErrorBoundary = withTranslation(NS_BRICK_KIT)(LegacyErrorBoundary);