+++ packages/editor-bricks-helper/src/EditorBrickAsComponent/EditorBrickAsComponent.spec.tsx
@@ -2,15 +2,24 @@ import React from "react";
 import { act } from "react-dom/test-utils";
 import { mount } from "enzyme";
 import { useDrag } from "react-dnd";
-import { BrickAsComponent } from "@next-core/brick-kit";
 import { UseSingleBrickConf } from "@next-core/brick-types";
 import { EditorBrickAsComponent } from "./EditorBrickAsComponent";
 import { getEditorBrick } from "./getEditorBrick";
 import { BuilderRuntimeNode, EditorSelfLayout } from "../interfaces";
+import { useBuilderData } from "../hooks/useBuilderData";
 
 jest.mock("react-dnd");
-jest.mock("@next-core/brick-kit");
+jest.mock("@next-core/brick-kit", () => ({
+  BrickAsComponent({ useBrick }: { useBrick: UseSingleBrickConf }) {
+    return (
+      <span>
+        BrickAsComponent({useBrick.brick},{useBrick.properties.nodeUid})
+      </span>
+    );
+  },
+}));
 jest.mock("./getEditorBrick");
+jest.mock("../hooks/useBuilderData");
 
 (useDrag as jest.MockedFunction<typeof useDrag>).mockReturnValue([
   { isDragging: false },
@@ -22,29 +31,22 @@ jest.mock("./getEditorBrick");
   typeof getEditorBrick
 >).mockResolvedValue("any-brick--editor");
 
-(BrickAsComponent as jest.MockedFunction<
-  typeof BrickAsComponent
->).mockImplementation(function MockBrickAsComponent({ useBrick }) {
-  return (
-    <span>
-      BrickAsComponent({(useBrick as UseSingleBrickConf).brick},
-      {(useBrick as UseSingleBrickConf).properties.nodeUid},
-      {(useBrick as UseSingleBrickConf).properties.brick})
-    </span>
-  );
-});
+const mockUseBuilderData = useBuilderData as jest.Mock;
 
 customElements.define(
   "any-brick--editor",
   class Tmp extends HTMLElement {
     static get selfLayout(): EditorSelfLayout {
-      return EditorSelfLayout.INLINE;
+      return undefined;
     }
   }
 );
 
 describe("EditorBrickAsComponent", () => {
   it("should work", async () => {
+    mockUseBuilderData.mockReturnValue({
+      edges: [],
+    });
     const wrapper = mount(
       <EditorBrickAsComponent
         node={
@@ -61,12 +63,39 @@ describe("EditorBrickAsComponent", () => {
       await (global as any).flushPromises();
     });
     wrapper.update();
-    expect(wrapper.text()).toBe(
-      "BrickAsComponent(any-brick--editor,1,any-brick)"
-    );
+    expect(wrapper.text()).toBe("BrickAsComponent(any-brick--editor,1)");
 
-    expect(wrapper.find("div").at(0).prop("className")).not.toContain(
-      "dragging"
+    const element = wrapper.find("div").at(0);
+    expect(element.hasClass("dragging")).toBe(false);
+    expect(element.hasClass("selfLayoutInline")).toBe(true);
+  });
+
+  it("should work for brick which has children", async () => {
+    mockUseBuilderData.mockReturnValue({
+      edges: [
+        {
+          parent: 1,
+        },
+      ],
+    });
+    const wrapper = mount(
+      <EditorBrickAsComponent
+        node={
+          ({
+            $$uid: 1,
+            instanceId: "instance-a",
+            id: "B-1",
+            brick: "any-brick",
+          } as Partial<BuilderRuntimeNode>) as BuilderRuntimeNode
+        }
+      />
     );
+    await act(async () => {
+      await (global as any).flushPromises();
+    });
+    wrapper.update();
+
+    const element = wrapper.find("div").at(0);
+    expect(element.hasClass("selfLayoutContainer")).toBe(true);
   });
 });