+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.5](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.5.4...@next-core/build-config-factory@2.5.5) (2021-03-24)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+
+
+
+
+## [2.5.4](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.5.3...@next-core/build-config-factory@2.5.4) (2021-03-18)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+
+
+
+
+## [2.5.3](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.5.2...@next-core/build-config-factory@2.5.3) (2021-03-17)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+
+
+
+
+## [2.5.2](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.5.1...@next-core/build-config-factory@2.5.2) (2021-03-17)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+
+
+
+
+## [2.5.1](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.5.0...@next-core/build-config-factory@2.5.1) (2021-03-16)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+
+
+
+
+# [2.5.0](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.4.8...@next-core/build-config-factory@2.5.0) (2021-03-12)
+
+
+### Features
+
+* remove unused api of syncDllDeps ([2250ede](https://github.com/easyops-cn/next-core/commit/2250edeb58190ec4ad3b885b4df88fec8977046d))
+
+
+
+
+
+## [2.4.8](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.4.7...@next-core/build-config-factory@2.4.8) (2021-03-08)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+
+
+
+
 ## [2.4.7](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.4.6...@next-core/build-config-factory@2.4.7) (2021-03-04)
 
 **Note:** Version bump only for package @next-core/build-config-factory