+++ packages/easyops-analytics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.8](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-analytics@0.1.7...@next-core/easyops-analytics@0.1.8) (2021-03-08)
+
+**Note:** Version bump only for package @next-core/easyops-analytics
+
+
+
+
+
 ## [0.1.7](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-analytics@0.1.6...@next-core/easyops-analytics@0.1.7) (2021-02-24)